+++ RL_DQN/test/include/globals.h
@@ -7,5 +7,5 @@ extern float GRAD_DECAY;
 extern int OPTIMIZER;
 
 //Network parameters
-extern std::vector<int> LAYER_SIZES;
+extern std::vector<size_t> LAYER_SIZES;
 extern std::vector<std::string> ACTIVATION_FUNCTIONS;
